bumping ketcher version, see base repo

+++ src/application/editor/tools/Clear.ts
@@ -13,33 +13,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
+
 import { CoreEditor, EditorHistory } from 'application/editor/internal';
-import { isSequenceMode } from 'application/editor/modes';
 import { BaseTool } from 'application/editor/tools/Tool';
 import { ReinitializeModeOperation } from 'application/editor/operations/modes';
 
 class ClearTool implements BaseTool {
-  private history: EditorHistory;
-
-  constructor(private editor: CoreEditor) {
+  constructor(private readonly editor: CoreEditor) {
     this.editor = editor;
-    this.history = new EditorHistory(editor);
+
+    // Only update history if there are entities to delete
+    if (!this.editor.drawingEntitiesManager.hasDrawingEntities) {
+      return;
+    }
+
+    const history = EditorHistory.getInstance(editor);
     const mode = editor.mode;
-    const isCurrentModeSequence = isSequenceMode(mode);
-    const isSequenceEditMode = isCurrentModeSequence && mode.isEditMode;
 
     const modelChanges = this.editor.drawingEntitiesManager.deleteAllEntities();
 
-    if (isCurrentModeSequence) {
+    if (mode.modeName === 'sequence-layout-mode') {
       modelChanges.addOperation(new ReinitializeModeOperation());
     }
 
+    this.editor.transientDrawingView.clear();
     this.editor.renderersContainer.update(modelChanges);
-    this.history.update(modelChanges);
-
-    if (isSequenceEditMode) {
-      mode.startNewSequence();
-    }
+    history.update(modelChanges);
   }
 
   destroy() {}
